backend/sentiment-analysis/requirements.txt
google-cloud-language==2.3.0
